backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
asyncpg
pydantic[email]>=2.0
passlib[bcrypt]
python-jose[cryptography]
openai
langchain==0.1.20
langchain-openai==0.1.7
langchain-community==0.0.38
langchain-core==0.1.53
numpy<2
easyocr
Pillow
python-multipart
aiofiles
